+++ R/file_report.R
@@ -30,7 +30,15 @@ file_report <- function(hash = NULL, ...) {
     res    <- virustotal_GET(path = "file/report", query = params, ...)
 
     if (res$response_code == 0 ){
-      res_df <- read.table(text = "", col.names = c("service", "detected", "version", "update", "result", "scan_id", "sha1", "resource", "response_code", "scan_date", "permalink", "verbose_msg", "total, positives", "sha256", "md5"))
+      res_df <- read.table(text = "", col.names = c("service", "detected",
+                                                    "version", "update",
+                                                    "result", "scan_id",
+                                                    "sha1", "resource",
+                                                    "response_code",
+                                                    "scan_date", "permalink",
+                                                    "verbose_msg",
+                                                    "total, positives",
+                                                    "sha256", "md5"))
       res_df[1, match(names(res), names(res_df))] <- res
       return(res_df)
     }