+++ R/rescan_file.R
@@ -35,7 +35,13 @@ rescan_file <- function(hash = NULL, ...) {
     res   <- virustotal_POST(path = "file/rescan", query = params, ...)
 
     if (res$response_code == 0 ){
-      res_df <- read.table(text = "", col.names = c("scans", "scan_id", "sha1", "resource", "response_code", "scan_date", "permalink", "verbose_msg", "total", "positives", "sha256", "md5"))
+      res_df <- read.table(text = "", col.names = c("scans", "scan_id", "sha1",
+                                                    "resource", "response_code",
+                                                     "scan_date", "permalink",
+                                                     "verbose_msg", "total",
+                                                     "positives", "sha256",
+                                                     "md5"))
+
       res_df[1, match(names(res), names(res_df))] <- res
       return(res_df)
     }