+++ R/url_report.R
@@ -9,8 +9,8 @@
 #' When scan is set to \code{1}, the result includes a \code{scan_id} field, which can be used again to retrieve the report. 
 #' @param \dots Additional arguments passed to \code{\link{virustotal_GET}}.
 #'  
-#' @return data.frame with 12 columns: 
-#' \code{scan_id, resource, url, response_code, scan_date, permalink, verbose_msg, positives, total, detected, result, detail}
+#' @return data.frame with 13 columns: 
+#' \code{scan_id, resource, url, response_code, scan_date, permalink, verbose_msg, positives, total, .id, detected, result, detail}
 #'  
 #' @export
 #' 
@@ -30,34 +30,41 @@ url_report <- function(url = NULL, scan_id = NULL, scan = 1, ...) {
 
   if (!is.character(url) & !is.character(scan_id)) {
         stop("Must specify url or scan_id.\n")
-    }
+  }
 
-    if (! (scan %in% c("0", "1"))) stop("scan must be either 0 or 1.\n")
+  if (! (scan %in% c("0", "1"))) stop("scan must be either 0 or 1.\n")
 
-    params <- list(resource = url, scan_id = scan_id, scan = scan)
+  params <- list(resource = url, scan_id = scan_id, scan = scan)
 
-    res    <- virustotal_POST(path = "url/report", query = params, ...)
+  res    <- virustotal_POST(path = "url/report", query = params, ...)
 
-    # Initialize empty data.frame
-    res_df <- read.table(text = "", col.names = c("scan_id", "resource", "url", "response_code", "scan_date", "permalink", "verbose_msg", "positives", "total", "detected", "result", "detail"))
+  # Initialize empty data.frame
+  res_df <- read.table(text = "", col.names = c("scan_id", "resource", "url",
+                                                "response_code", "scan_date",
+                                                "permalink", "verbose_msg",
+                                                "positives", "total",
+                                                "detected", "result",
+                                                "detail"))
 
-    if ( !is.null(scan_id) & length(res) == 0) {
-      warning("No results returned. Likely cause: incorrect scan_id.\n")
-      res_df[1, "scan_id"] <- scan_id
+  if ( !is.null(scan_id) & length(res) == 0) {
+    warning("No results returned. Likely cause: incorrect scan_id.\n")
+    res_df[1, "scan_id"] <- scan_id
+    return(res_df)
+
+  } else if (res$response_code == 0) {
+    warning("No reports for the URL available. Set scan to 1 to submit URL
+                                                               for scanning.\n")
+    res_df[1, match(names(res), names(res_df))] <- res
+    return(res_df)
+
+  } else if (!is.null(url) & length(res) < 11) {
+      warning("No reports for the URL available. The URL has been successfully
+                        submitted for scanning. Come back later for results.\n")
+      res_df[1, match(names(res), names(res_df))] <- res
       return(res_df)
-
-    } else if (res$response_code == 0) {
-      warning("No reports for the URL available. Set scan to 1 to submit URL for scanning.\n")
-     res_df[1, match(names(res), names(res_df))] <- res
-     return(res_df)
-
-    } else if (!is.null(url) & length(res) < 11) {
-        warning("No reports for the URL available. The URL has been successfully submitted for scanning. Come back later for results.\n")
-        res_df[1, match(names(res), names(res_df))] <- res
-        return(res_df)
     }
 
     res_10 <- do.call(cbind, lapply(res[1:10], unlist))
-    res_11 <- do.call(rbind.fill, lapply(res$scans, as.data.frame))
+    res_11 <- ldply(res$scans, as.data.frame)
     cbind(res_10, res_11)
 }