+++ R/virustotal.R
@@ -28,17 +28,19 @@ NULL
 #' @param \dots Additional arguments passed to \code{\link[httr]{GET}}.
 #' @return list
 
-virustotal_GET <- function(query=list(), path = path, key = Sys.getenv("VirustotalToken"), ...) {
+virustotal_GET <- function(query=list(), path = path,
+                                     key = Sys.getenv("VirustotalToken"), ...) {
 
   if (identical(key, "")) {
-        stop("Please set application id and password using set_key(key='key')).\n")
+        stop("Please set application key via set_key(key='key')).\n")
   }
 
   query$apikey <- key
 
   rate_limit()
 
-  res <- GET("http://www.virustotal.com/", path = paste0("vtapi/v2/", path), query = query, ...)
+  res <- GET("http://www.virustotal.com/", path = paste0("vtapi/v2/", path),
+                                                             query = query, ...)
   virustotal_check(res)
   res <- content(res)
 
@@ -55,17 +57,19 @@ virustotal_GET <- function(query=list(), path = path, key = Sys.getenv("Virustot
 #' @param \dots Additional arguments passed to \code{\link[httr]{POST}}.
 #' @return list
 
-virustotal_POST <- function(query=list(), path = path, body=NULL, key = Sys.getenv("VirustotalToken"), ...) {
+virustotal_POST <- function(query=list(), path = path, body=NULL,
+                                     key = Sys.getenv("VirustotalToken"), ...) {
 
   if (identical(key, "")) {
-        stop("Please set application id and password using set_key(key='key')).\n")
+        stop("Please set application key via set_key(key='key')).\n")
   }
 
   query$apikey <- key
 
   rate_limit()
 
-  res <- POST("http://www.virustotal.com/", path = paste0("vtapi/v2/", path), query = query, body = body, ...)
+  res <- POST("http://www.virustotal.com/", path = paste0("vtapi/v2/", path),
+                                                query = query, body = body, ...)
   virustotal_check(res)
   res <- content(res)
 
@@ -80,7 +84,8 @@ virustotal_POST <- function(query=list(), path = path, body=NULL, key = Sys.gete
 
 virustotal_check <- function(req) {
 
-  if (req$status_code == 204) stop("Rate Limit Exceeded. Only 4 Queries per minute allowed.\n")
+  if (req$status_code == 204) stop("Rate Limit Exceeded.
+                                          Only 4 Queries per minute allowed.\n")
   if (req$status_code < 400) return(invisible())
 
   stop("HTTP failure: ", req$status_code, "\n", call. = FALSE)
@@ -89,21 +94,27 @@ virustotal_check <- function(req) {
 #' 
 #' Rate Limits
 #' 
-#' Virustotal requests throttled at 4 per minute. This function creates an env. variable 
+#' Virustotal requests throttled at 4 per min. The function creates an env. var.
 #' that tracks number of requests per minute, and enforces appropriate waiting.
 #' 
 
 rate_limit <- function() {
 
   # First request --- initialize time of first request and request count
-  if (Sys.getenv("VT_RATE_LIMIT") == "") return(Sys.setenv(VT_RATE_LIMIT = paste0(0, ",", Sys.time(), ",", 0)))
+  if (Sys.getenv("VT_RATE_LIMIT") == "") {
+    return(Sys.setenv(VT_RATE_LIMIT = paste0(0, ",", Sys.time(), ",", 0)))
+  }
 
   rate_lim         <- Sys.getenv("VT_RATE_LIMIT")
   req_count        <- as.numeric(gsub(",.*", "", rate_lim)) + 1
-  past_duration    <- as.numeric(strsplit(rate_lim, ",")[[1]][3], units = "secs")
-  current_duration <- difftime(Sys.time(), as.POSIXct(strsplit(rate_lim, ",")[[1]][2]), units = "secs")
+  past_duration    <- as.numeric(strsplit(rate_lim, ",")[[1]][3],
+                                                                 units = "secs")
+  current_duration <- difftime(Sys.time(),
+                    as.POSIXct(strsplit(rate_lim, ",")[[1]][2]), units = "secs")
 
-  if (current_duration > 60) return(Sys.setenv(VT_RATE_LIMIT = paste0(1, ",", Sys.time(), ",", 0)))
+  if (current_duration > 60) {
+    return(Sys.setenv(VT_RATE_LIMIT = paste0(1, ",", Sys.time(), ",", 0)))
+  }
 
   net_duration     <- past_duration + current_duration
 
@@ -113,5 +124,6 @@ rate_limit <- function() {
     return(Sys.setenv(VT_RATE_LIMIT = paste0(1, ",", Sys.time(), ",", 0)))
   }
 
-  return(Sys.setenv(VT_RATE_LIMIT = paste0(req_count, ",", Sys.time(), ",", net_duration)))
+  return(Sys.setenv(VT_RATE_LIMIT =
+                         paste0(req_count, ",", Sys.time(), ",", net_duration)))
 }