+++ NEWS.md
@@ -1,5 +1,6 @@
 # virustotal 0.2.0
 
+* Removed link to bitdefender because CRAN was having issues
 * Better documentation with examples including comment for set_key, better formatting
 * Better error handling and more consistent returned data structures for url_report, file_report, rescan_file 
 * url_report now accepts scan_id as a param