+++ NEWS.md
@@ -8,7 +8,8 @@
 * Added more tests, specifically checking returns to what happens when params/hash are incorrect  
 * Enforces rate limiting --- 4 queries per minute. 
 * Graceful error handling if error limit exceeded.
-
+* changed virustotal to VirusTotal as CRAN doesn't muck around.
+ 
 # virustotal 0.1.0
 
 * Initial release